refactor(redskilled): extract client rendezvous (#3481)
